+++ src/main/scala/algorithm/tree/Node.scala
@@ -4,5 +4,5 @@ package algorithm.tree
 import model.DataPoint
 
 trait Node {
-    def predict(dataPoint: DataPoint): Int
+    def predict(dataPoint: DataPoint): Double
 }
